requirements.txt
torch==2.1.0
torchvision
numpy
tqdm
tensorboard
matplotlib
seaborn
pandas

speechbrain~=0.5.12
datasets
torch_audiomentations
editdistance
wandb
pyctcdecode
torchaudio
pillow
notebook==6.5.6
sentencepiece
flashlight-text
scipy
Unidecode
inflect
pyworld
